modularized p4 and weapons
